requirements.txt
opencv-python-headless
pyautogui
pillow
pynput
pyqt5
numpy